refactor: board comment entity one to many bidirectional relationship change one to many unidirectional relationship

+++ src/backend/community-service/src/main/java/harmony/communityservice/community/query/dto/SearchBoardResponse.java
@@ -15,6 +15,6 @@ public record SearchBoardResponse(
         ModifiedType modified,
         String createdAt,
         List<SearchEmojiResponse> searchEmojiResponses,
-        int commentCount
+        Long commentCount
 ) {
 }
